Refactor code

+++ src/main/java/ch/admin/bit/eid/verifier_management/models/entities/VerificationRequestObject.java
@@ -32,5 +32,5 @@ public class VerificationRequestObject {
     private ClientMetadata clientMetadata;
 
     @TimeToLive
-    private int expiresAt;
+    private long expiresAt;
 }
